FillMaskPipeline
